+++ TJC.Persist.Tests/Mocks/ExamplePersistObject.cs
@@ -1,0 +1,7 @@
+﻿namespace TJC.Persist.Tests.Mocks;
+
+internal class ExamplePersistObject : IPersistObject
+{
+    public string? Name { get; set; }
+    public int Value { get; set; }
+}