+++ TJC.Persist/IPersistObject.cs
@@ -1,0 +1,8 @@
+﻿namespace TJC.Persist;
+
+/// <summary>
+/// Allow object to use <see cref="PersistManager"/> methods
+/// </summary>
+public interface IPersistObject
+{
+}